+++ dialect/mysql/dialect/select.go
@@ -29,6 +29,7 @@ type SelectQuery struct {
 	clause.Offset
 	clause.For
 	bob.Load
+	bob.ContextualModdable[*SelectQuery]
 }
 
 func (s *SelectQuery) SetInto(i any) {
@@ -38,6 +39,10 @@ func (s *SelectQuery) SetInto(i any) {
 func (s SelectQuery) WriteSQL(ctx context.Context, w io.Writer, d bob.Dialect, start int) ([]any, error) {
 	var args []any
 	var err error
+
+	if ctx, err = s.RunContextualMods(ctx, &s); err != nil {
+		return nil, err
+	}
 
 	withArgs, err := bob.ExpressIf(ctx, w, d, start+len(args), s.With,
 		len(s.With.CTEs) > 0, "\n", "")