+++ dialect/mysql/table.go
@@ -68,21 +68,21 @@ type Table[T orm.Table, Tslice ~[]T, Tset setter[T]] struct {
 	pkExpr        dialect.Expression
 	setterMapping mappings.Mapping
 
-	BeforeInsertHooks orm.Hooks[[]Tset, orm.SkipModelHooksKey]
-	AfterInsertHooks  orm.Hooks[Tslice, orm.SkipModelHooksKey]
+	BeforeInsertHooks bob.Hooks[[]Tset, bob.SkipModelHooksKey]
+	AfterInsertHooks  bob.Hooks[Tslice, bob.SkipModelHooksKey]
 
-	BeforeUpsertHooks orm.Hooks[[]Tset, orm.SkipModelHooksKey]
-	AfterUpsertHooks  orm.Hooks[Tslice, orm.SkipModelHooksKey]
+	BeforeUpsertHooks bob.Hooks[[]Tset, bob.SkipModelHooksKey]
+	AfterUpsertHooks  bob.Hooks[Tslice, bob.SkipModelHooksKey]
 
-	BeforeUpdateHooks orm.Hooks[Tslice, orm.SkipModelHooksKey]
-	AfterUpdateHooks  orm.Hooks[Tslice, orm.SkipModelHooksKey]
+	BeforeUpdateHooks bob.Hooks[Tslice, bob.SkipModelHooksKey]
+	AfterUpdateHooks  bob.Hooks[Tslice, bob.SkipModelHooksKey]
 
-	BeforeDeleteHooks orm.Hooks[Tslice, orm.SkipModelHooksKey]
-	AfterDeleteHooks  orm.Hooks[Tslice, orm.SkipModelHooksKey]
+	BeforeDeleteHooks bob.Hooks[Tslice, bob.SkipModelHooksKey]
+	AfterDeleteHooks  bob.Hooks[Tslice, bob.SkipModelHooksKey]
 
-	InsertQueryHooks orm.Hooks[*dialect.InsertQuery, orm.SkipQueryHooksKey]
-	UpdateQueryHooks orm.Hooks[*dialect.UpdateQuery, orm.SkipQueryHooksKey]
-	DeleteQueryHooks orm.Hooks[*dialect.DeleteQuery, orm.SkipQueryHooksKey]
+	InsertQueryHooks bob.Hooks[*dialect.InsertQuery, bob.SkipQueryHooksKey]
+	UpdateQueryHooks bob.Hooks[*dialect.UpdateQuery, bob.SkipQueryHooksKey]
+	DeleteQueryHooks bob.Hooks[*dialect.DeleteQuery, bob.SkipQueryHooksKey]
 
 	// The AUTO_INCREMENT column that we can use to retrieve values using lastInsertID
 	// If empty, there is no auto inc
@@ -153,7 +153,7 @@ func (t *Table[T, Tslice, Tset]) InsertMany(ctx context.Context, exec bob.Execut
 
 	var err error
 
-	ctx, err = t.BeforeInsertHooks.Do(ctx, exec, rows)
+	ctx, err = t.BeforeInsertHooks.RunHooks(ctx, exec, rows)
 	if err != nil {
 		return nil, err
 	}
@@ -163,7 +163,7 @@ func (t *Table[T, Tslice, Tset]) InsertMany(ctx context.Context, exec bob.Execut
 	)
 
 	// To prevent unnecessary work, we will do this before we add the rows
-	ctx, err = t.InsertQueryHooks.Do(ctx, exec, q.Expression)
+	ctx, err = t.InsertQueryHooks.RunHooks(ctx, exec, q.Expression)
 	if err != nil {
 		return nil, err
 	}
@@ -194,7 +194,7 @@ func (t *Table[T, Tslice, Tset]) InsertMany(ctx context.Context, exec bob.Execut
 		}
 	}
 
-	_, err = t.AfterInsertHooks.Do(ctx, exec, inserted)
+	_, err = t.AfterInsertHooks.RunHooks(ctx, exec, inserted)
 	if err != nil {
 		return nil, err
 	}
@@ -210,7 +210,7 @@ func (t *Table[T, Tslice, Tset]) Update(ctx context.Context, exec bob.Executor, 
 		return nil
 	}
 
-	_, err := t.BeforeUpdateHooks.Do(ctx, exec, rows)
+	_, err := t.BeforeUpdateHooks.RunHooks(ctx, exec, rows)
 	if err != nil {
 		return err
 	}
@@ -222,7 +222,7 @@ func (t *Table[T, Tslice, Tset]) Update(ctx context.Context, exec bob.Executor, 
 
 	q := Update(um.Table(t.NameAs()), vals, um.Where(t.pkExpr.In(pkPairs...)))
 
-	ctx, err = t.UpdateQueryHooks.Do(ctx, exec, q.Expression)
+	ctx, err = t.UpdateQueryHooks.RunHooks(ctx, exec, q.Expression)
 	if err != nil {
 		return err
 	}
@@ -235,7 +235,7 @@ func (t *Table[T, Tslice, Tset]) Update(ctx context.Context, exec bob.Executor, 
 		vals.Overwrite(row)
 	}
 
-	if _, err = t.AfterUpdateHooks.Do(ctx, exec, rows); err != nil {
+	if _, err = t.AfterUpdateHooks.RunHooks(ctx, exec, rows); err != nil {
 		return err
 	}
 
@@ -268,7 +268,7 @@ func (t *Table[T, Tslice, Tset]) UpsertMany(ctx context.Context, exec bob.Execut
 
 	var err error
 
-	ctx, err = t.BeforeUpsertHooks.Do(ctx, exec, rows)
+	ctx, err = t.BeforeUpsertHooks.RunHooks(ctx, exec, rows)
 	if err != nil {
 		return nil, err
 	}
@@ -293,7 +293,7 @@ func (t *Table[T, Tslice, Tset]) UpsertMany(ctx context.Context, exec bob.Execut
 	)
 
 	// To prevent unnecessary work, we will do this before we add the rows
-	ctx, err = t.InsertQueryHooks.Do(ctx, exec, q.Expression)
+	ctx, err = t.InsertQueryHooks.RunHooks(ctx, exec, q.Expression)
 	if err != nil {
 		return nil, err
 	}
@@ -325,7 +325,7 @@ func (t *Table[T, Tslice, Tset]) UpsertMany(ctx context.Context, exec bob.Execut
 		}
 	}
 
-	_, err = t.AfterUpsertHooks.Do(ctx, exec, upserted)
+	_, err = t.AfterUpsertHooks.RunHooks(ctx, exec, upserted)
 	if err != nil {
 		return nil, err
 	}
@@ -340,7 +340,7 @@ func (t *Table[T, Tslice, Tset]) Delete(ctx context.Context, exec bob.Executor, 
 		return nil
 	}
 
-	_, err := t.BeforeDeleteHooks.Do(ctx, exec, rows)
+	_, err := t.BeforeDeleteHooks.RunHooks(ctx, exec, rows)
 	if err != nil {
 		return err
 	}
@@ -352,7 +352,7 @@ func (t *Table[T, Tslice, Tset]) Delete(ctx context.Context, exec bob.Executor, 
 
 	q := Delete(dm.From(t.NameAs()), dm.Where(t.pkExpr.In(pkPairs...)))
 
-	ctx, err = t.DeleteQueryHooks.Do(ctx, exec, q.Expression)
+	ctx, err = t.DeleteQueryHooks.RunHooks(ctx, exec, q.Expression)
 	if err != nil {
 		return err
 	}
@@ -361,7 +361,7 @@ func (t *Table[T, Tslice, Tset]) Delete(ctx context.Context, exec bob.Executor, 
 		return err
 	}
 
-	if _, err = t.AfterDeleteHooks.Do(ctx, exec, rows); err != nil {
+	if _, err = t.AfterDeleteHooks.RunHooks(ctx, exec, rows); err != nil {
 		return err
 	}
 
@@ -479,14 +479,14 @@ type TQuery[Q bob.Expression, T any, Ts ~[]T] struct {
 	ctx   context.Context
 	exec  bob.Executor
 	view  *View[T, Ts]
-	hooks *orm.Hooks[Q, orm.SkipQueryHooksKey]
+	hooks *bob.Hooks[Q, bob.SkipQueryHooksKey]
 }
 
 // Execute the query
 func (t *TQuery[Q, T, Tslice]) Exec() (int64, error) {
 	var err error
 
-	if t.ctx, err = t.hooks.Do(t.ctx, t.exec, t.Expression); err != nil {
+	if t.ctx, err = t.hooks.RunHooks(t.ctx, t.exec, t.Expression); err != nil {
 		return 0, err
 	}