+++ dialect/sqlite/fm/qm.go
@@ -4,11 +4,10 @@ import (
 	"github.com/stephenafamo/bob"
 	"github.com/stephenafamo/bob/clause"
 	"github.com/stephenafamo/bob/dialect/sqlite/dialect"
-	"github.com/stephenafamo/bob/mods"
 )
 
 func Distinct() bob.Mod[*dialect.Function] {
-	return mods.QueryModFunc[*dialect.Function](func(f *dialect.Function) {
+	return bob.ModFunc[*dialect.Function](func(f *dialect.Function) {
 		f.Distinct = true
 	})
 }
@@ -22,7 +21,7 @@ func OrderBy(e any) dialect.OrderBy[*dialect.Function] {
 }
 
 func Filter(e ...any) bob.Mod[*dialect.Function] {
-	return mods.QueryModFunc[*dialect.Function](func(f *dialect.Function) {
+	return bob.ModFunc[*dialect.Function](func(f *dialect.Function) {
 		f.Filter = append(f.Filter, e...)
 	})
 }