+++ hooks_test.go
@@ -1,4 +1,4 @@
-package orm
+package bob
 
 import (
 	"context"
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/stephenafamo/bob"
 )
 
 func TestHooks(t *testing.T) {
@@ -16,18 +15,18 @@ func TestHooks(t *testing.T) {
 	// Test Adding Hooks
 	for i := 0; i < 5; i++ {
 		initial := len(H.hooks)
-		f := func(ctx context.Context, _ bob.Executor, s *string) (context.Context, error) {
+		f := func(ctx context.Context, _ Executor, s *string) (context.Context, error) {
 			*s = *s + fmt.Sprintf("%d", initial+1)
 			return ctx, nil
 		}
-		H.Add(f)
+		H.AppendHooks(f)
 		if len(H.hooks) != initial+1 {
 			t.Fatalf("Did not add hook number %d", i+1)
 		}
 	}
 
 	s := ""
-	if _, err := H.Do(context.Background(), nil, &s); err != nil {
+	if _, err := H.RunHooks(context.Background(), nil, &s); err != nil {
 		t.Fatal(err)
 	}
 	if diff := cmp.Diff("12345", s); diff != "" {
@@ -36,7 +35,7 @@ func TestHooks(t *testing.T) {
 
 	// test skipping hooks
 	s = ""
-	if _, err := H.Do(context.WithValue(context.Background(), skipKey{}, true), nil, &s); err != nil {
+	if _, err := H.RunHooks(context.WithValue(context.Background(), skipKey{}, true), nil, &s); err != nil {
 		t.Fatal(err)
 	}
 	if diff := cmp.Diff("", s); diff != "" {