+++ mods/mods.go
@@ -17,12 +17,6 @@ func (q QueryMods[T]) Apply(query T) {
 	}
 }
 
-type QueryModFunc[T any] func(T)
-
-func (q QueryModFunc[T]) Apply(query T) {
-	q(query)
-}
-
 // This is a generic type for expressions can take extra mods as a function
 // allows for some fluent API, for example with functions
 type Moddable[T bob.Expression] func(...bob.Mod[T]) T
@@ -175,4 +169,10 @@ type set[Q interface{ AppendSet(clauses ...any) }] []any
 
 func (s set[Q]) Apply(q Q) {
 	q.AppendSet(s...)
+}
+
+type Hook[Q interface{ AppendHooks(...bob.Hook[Q]) }] []bob.Hook[Q]
+
+func (h Hook[Q]) Apply(q Q) {
+	q.AppendHooks(h...)
 }